+++ src/main/java/kitchenpos/domain/OrderLineItems.java
@@ -1,0 +1,37 @@
+package kitchenpos.domain;
+
+import org.springframework.util.CollectionUtils;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.FetchType;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
+
+@Embeddable
+public class OrderLineItems {
+
+    @OneToMany(fetch = FetchType.EAGER, cascade = CascadeType.PERSIST)
+    @JoinColumn(name = "order_id")
+    private List<OrderLineItem> orderLineItems;
+
+    protected OrderLineItems() {
+    }
+
+    public OrderLineItems(List<OrderLineItem> orderLineItems) {
+        validateOrderLineItems(orderLineItems);
+        this.orderLineItems = new ArrayList<>(orderLineItems);
+    }
+
+    private void validateOrderLineItems(List<OrderLineItem> orderLineItems) {
+        if (CollectionUtils.isEmpty(orderLineItems)) {
+            throw new IllegalArgumentException("주문 항목은 하나 이상이여야 합니다");
+        }
+    }
+
+    public List<OrderLineItem> getItems() {
+        return orderLineItems;
+    }
+}
